Remove unused old test files, organize imports and fix server-client sending message bug
