+++ sourceCode/src/com/stamacoding/rsaApp/log/debug/Debug.java
@@ -2,9 +2,7 @@ package com.stamacoding.rsaApp.log.debug;
 
 import java.time.OffsetDateTime;
 
-import com.stamacoding.rsaApp.log.filesystem.Filesystem;
 import com.stamacoding.rsaApp.log.logger.Logger;
-import com.stamacoding.rsaApp.log.filesystem.Filesystem;
 
 public class Debug
 {
