+++ sourceCode/src/com/stamacoding/rsaApp/log/logger/Logger.java
@@ -1,12 +1,8 @@
 package com.stamacoding.rsaApp.log.logger;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.Formatter;
-import java.util.Scanner;
 
 import com.stamacoding.rsaApp.log.debug.Debug;
