--- sourceCode/src/com/stamacoding/rsaApp/server/Client.java
@@ -1,5 +1,0 @@
-package com.stamacoding.rsaApp.server;
-
-public class Client {
-
-}
