--- testCode/.settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/server/TestServer.java=UTF-8
