--- testCode/logFile.txt
@@ -1,17 +1,0 @@
-
-[2020-06-13T19:03:52.007163700+02:00][RaspberryPi][DEBUG]: [Trying to connect to raspberry pi (77.21.46.46:1001)]
-[2020-06-13T19:03:52.033148+02:00][RaspberryPi][DEBUG]: [Connected to raspberry pi (77.21.46.46:1001)]
-[2020-06-13T19:03:52.034147500+02:00][RaspberryPi][DEBUG]: [Result: SUCCESS]
-[2020-06-13T19:06:22.399286100+02:00][RaspberryPi][DEBUG]: [Trying to connect to raspberry pi (77.21.46.46:1001)]
-[2020-06-13T19:06:22.413265500+02:00][RaspberryPi][DEBUG]: [Connected to raspberry pi (77.21.46.46:1001)]
-[2020-06-13T19:06:22.414264300+02:00][RaspberryPi][DEBUG]: [Result: SUCCESS]
-[2020-06-13T19:09:49.307488800+02:00][RaspberryPi][DEBUG]: [Starting server application]
-[2020-06-13T19:09:49.321481300+02:00][RaspberryPi][DEBUG]: [Server application running]
-[2020-06-13T19:10:00.030468300+02:00][RaspberryPi][DEBUG]: [Starting server application]
-[2020-06-13T19:10:00.043461100+02:00][RaspberryPi][DEBUG]: [Server application running]
-[2020-06-13T19:12:10.985847600+02:00][RaspberryPi][DEBUG]: [Starting server application]
-[2020-06-13T19:12:10.998841300+02:00][RaspberryPi][DEBUG]: [Server application running]
-[2020-06-13T19:12:14.872006300+02:00][RaspberryPi][DEBUG]: [Starting server application]
-[2020-06-13T19:12:14.886998600+02:00][RaspberryPi][DEBUG]: [Server application running]
-[2020-06-13T20:38:29.195006200+02:00][RaspberryPi][DEBUG]: [Starting server application]
-[2020-06-13T20:38:29.210996800+02:00][RaspberryPi][DEBUG]: [Server application running]
