--- testCode/src/server/New/TestClass.java
@@ -1,9 +1,0 @@
-package server.New;
-
-public class TestClass {
-
-	public static void main(String[] args) {
-		Client.run();
-	}
-
-}
