--- testCode/src/server/New/TestClassServer.java
@@ -1,7 +1,0 @@
-package server.New;
-
-public class TestClassServer {
-	public static void main(String[] args) {
-		Server.run();
-	}
-}
